refactor: emotion type 분리
