+++ apps/web/src/features/select-purchase-emotion/model/index.ts
@@ -1,0 +1,3 @@
+export type { Emotion, EmotionValue } from './types';
+
+export { emotions } from './emotion';
