+++ apps/web/src/features/select-purchase-emotion/model/types.ts
@@ -1,0 +1,14 @@
+export type EmotionValue =
+  | 'stress'
+  | 'impulse'
+  | 'survival'
+  | 'habit'
+  | 'self-reward'
+  | 'investment';
+
+export interface Emotion {
+  label: string;
+  value: EmotionValue;
+  tags: readonly string[];
+  description: string;
+}
