+++ apps/web/src/features/select-purchase-emotion/ui/ScrollPicker.tsx
@@ -2,7 +2,7 @@
 
 import React, { useRef, useEffect, useCallback } from 'react';
 import { Text } from '@/shared/ui/text/Text';
-import { emotions } from '../model/emotion';
+import { emotions } from '../model';
 import * as styles from './ScrollPicker.css';
 
 interface ScrollPickerProps {
