+++ apps/web/src/features/select-purchase-emotion/ui/index.ts
@@ -1,4 +1,5 @@
-export { DialWheel } from './DialWheel';
-export { ScrollPicker } from './ScrollPicker';
 export { EmotionSelector } from './EmotionSelector';
 export { EmotionDescription } from './EmotionDescription';
+
+export { DialWheel } from './DialWheel';
+export { ScrollPicker } from './ScrollPicker';
